GĐ 7-2: Hoàn thiện giao diện và logic xử lý file; cập nhật các view liên quan, đồng thời điều chỉnh controller và model của Task và Comment để tích hợp chức năng này.

+++ routes/web/comments.php
@@ -9,7 +9,7 @@ Route::middleware(['auth', 'verified'])
     ->group(function () {
         Route::get('/', 'index')->name('index');
 
-        Route::delete('{comment}/force-delete', 'forceDelete')->middleware('role:admin|super-admin')->withTrashed()->name('force-delete');
+        Route::delete('{comment}/force-delete', 'forceDelete')->middleware('role:admin|super-admin', 'password.confirm')->withTrashed()->name('force-delete');
 
         Route::get('recycle', 'recycle')->name('recycle');
         Route::post('{comment}/restore', 'restore')->withTrashed()->name('restore');
